--- model/leaderOrm.js
@@ -1,10 +1,0 @@
-const { selectTopLeaders } = require("./leaderQueries");
-const connection = require("../config/connection");
-
-const selectLeadersFromDb = async (top) => {
-  return await connection.query(selectTopLeaders, top);
-};
-
-module.exports = {
-  selectLeadersFromDb,
-};
