+++ routes/apiRoutes/index.js
@@ -1,9 +1,9 @@
 const router = require("express").Router();
-const leaderRoutes = require("./leaderRoutes");
+const scoreRoutes = require("./scoreRoutes");
 const userRoutes = require("./userRoutes");
 const generateWordRoutes = require("./generateWordRoutes");
 
-router.use("/leaderboard", leaderRoutes);
+router.use("/scores", scoreRoutes);
 router.use("/users", userRoutes);
 
 // Contains methods to generate random words
