+++ routes/apiRoutes/scoreRoutes/index.js
@@ -1,8 +1,11 @@
 const router = require("express").Router();
-const { getLeaderboard } = require("../../../controllers/leaderController");
+const {
+  getLeaderboard,
+  postScore,
+} = require("../../../controllers/scoreController");
 const authMiddleware = require("../../../middlewares/authorizationMiddleware");
 
 router.use(authMiddleware);
-router.route("/").get(getLeaderboard);
+router.route("/").get(getLeaderboard).post(postScore);
 
 module.exports = router;
